single module for pandas backend
