--- hm_fashion_recs/recs_utils.py
@@ -1,9 +1,0 @@
-import time
-from contextlib import contextmanager
-
-
-@contextmanager
-def timer(name):
-    start_time = time.time()
-    yield
-    print(f"[{name}] {time.time() - start_time:.3f} s")
