+++ utils/timer.py
@@ -1,6 +1,6 @@
 import time
 import logging
-from typing import Callable
+from typing import Callable, Iterator
 
 __all__ = ["TimerManager"]
 
@@ -9,6 +9,8 @@ logger = logging.getLogger(__name__)
 
 
 class Timer:
+    """Utility timer for TimerManager."""
+
     def __init__(self, report: Callable[[float], None]) -> None:
         self.report = report
         self.start_time = time.perf_counter()
@@ -24,34 +26,51 @@ class Timer:
 
 
 class TimerManager:
+    """Utility timer that can measure time using `timeit` function. Intended use is through context manager like
+    >>> tm = TimerManager
+    >>> with tm.timeit('heavy_call'):
+    >>>     heavy_call()
+    TimeManager supports nested timings if called through the same object.
+    """
+
     SEPARATOR = "."
 
-    def __init__(self, allow_overwrite=False, verbose=False) -> None:
+    def __init__(self, allow_overwrite=False) -> None:
         self.stack = []
         self.name2time = {}
         self.allow_overwrite = allow_overwrite
-        self.verbose = verbose
-
-    def report_timer(self, name, time):
-        if not self.allow_overwrite:
-            assert name not in self.name2time, f"Trying to rewrite measurment for {name}"
-        if self.verbose:
-            logger.info("%s time: %s", name, time)
-
-        self.name2time[name] = time
-
-    def timeit(self, name):
-        assert self.SEPARATOR not in name
-
-        self.stack.append(name)
-
-        name = self.SEPARATOR.join(self.stack)
-
-        def report(time):
-            self.report_timer(name, time)
-            self.stack.pop()
-
-        return Timer(report)
 
     def get_results(self):
         return dict(self.name2time)
+
+    def timeit(self, name):
+        self._validate_name(name)
+        self._push(name)
+
+        full_name = self._get_full_name()
+
+        def report(time):
+            self.report_timer(full_name, time)
+            self._pop()
+
+        return Timer(report)
+
+    def report_timer(self, name, time):
+        """Record timer result for a timer"""
+        if not self.allow_overwrite:
+            assert name not in self.name2time, f"Trying to rewrite measurment for {name}"
+        logger.info("%s time: %s", name, time)
+
+        self.name2time[name] = time
+
+    def _push(self, name):
+        self.stack.append(name)
+
+    def _pop(self):
+        self.stack.pop()
+
+    def _validate_name(self, name):
+        assert self.SEPARATOR not in name
+
+    def _get_full_name(self):
+        return self.SEPARATOR.join(self.stack)
